requirements.txt
boto3==1.18.53
botocore==1.21.53
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
marshmallow==3.13.0
mongoengine==0.23.1
mongomock==3.23.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
s3transfer==0.5.0
sentinels==1.0.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
